requirements.txt
fastapi
uvicorn
requests
passlib[argon2]
python-jose
python-multipart
prometheus-fastapi-instrumentator
slowapi
jinja2
